adiciona navigation e corrige bug de exibição
Acoontece que ao tentar realizar uma instanciação segura de _binding para binding, o layout não era carregado e ocasionava em bug.
