+++ Calculadora/.idea/misc.xml
@@ -3,8 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout-land/fragment_calculator.xml" value="0.16302083333333334" />
         <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout-port/activity_main.xml" value="0.20416666666666666" />
-        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/activity_main.xml" value="0.25" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout-port/fragment_calculator.xml" value="0.22395833333333334" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/activity_main.xml" value="0.16302083333333334" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/fragment_blank.xml" value="0.20416666666666666" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/fragment_calculator.xml" value="0.2390625" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/main_activity.xml" value="0.1" />
       </map>
     </option>
   </component>
