+++ Calculadora/app/src/main/AndroidManifest.xml
@@ -9,8 +9,7 @@
         android:roundIcon="@mipmap/ic_new_icon_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Calculadora">
-        <activity android:name=".MainActivity"
-            android:screenOrientation="portrait">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
