+++ Calculadora/app/src/main/java/gusoliveira21/calculadora/CalculatorViewModel.kt
@@ -1,15 +1,13 @@
 package gusoliveira21.calculadora
 
 import androidx.lifecycle.ViewModel
-import gusoliveira21.calculadora.databinding.ActivityMainBinding
+import gusoliveira21.calculadora.databinding.FragmentCalculatorBinding
 import net.objecthunter.exp4j.ExpressionBuilder
 
-class MainViewModel: ViewModel() {
+
+class CalculatorViewModel: ViewModel() {
     val simbol = listOf("+","-","/","*",".")
     val clean = ""
-
-
-
 
     fun onVerificaSeJaFoiDigitadoUmPontoAnteriormente(ViewCampoDigitado: CharSequence?): Boolean {
         val listaDeValores = listOf("1", "2", "3", "4", "5", "6", "7", "8", "9", "0")
@@ -72,7 +70,7 @@ class MainViewModel: ViewModel() {
         return ""
     }
 
-    fun onConditionalTest(binding: ActivityMainBinding):Boolean {
+    fun onConditionalTest(binding: FragmentCalculatorBinding):Boolean {
         if ((onCheckIfIsEmpty(binding.tvEntradaDados.text) || onCheckIfLastElementOfListIsSymbol(binding.tvEntradaDados.text)))
             return true
         return false
