+++ Calculadora/app/src/main/res/layout-port/fragment_calculator.xml
@@ -4,6 +4,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
+
+<androidx.constraintlayout.widget.ConstraintLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -12,16 +17,14 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
-        <!--viewGroup interior-->
         <androidx.appcompat.widget.LinearLayoutCompat
             android:id="@+id/box"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:background="@drawable/background"
             android:orientation="vertical"
             tools:context=".MainActivity">
 
-            <!--Toolbar exibe titulo-->
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/toolbar"
                 android:layout_width="match_parent"
@@ -79,16 +82,6 @@
                         tools:text="Entrada" />
                 </androidx.constraintlayout.widget.ConstraintLayout>
             </androidx.cardview.widget.CardView>
-
-
-            <!--LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"
-                tools:ignore="MissingConstraints">
-
-
-            </LinearLayout-->
 
 
             <LinearLayout
@@ -260,4 +253,5 @@
             </LinearLayout>
         </androidx.appcompat.widget.LinearLayoutCompat>
     </ScrollView>
+</androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
