+++ Calculadora/app/src/main/res/values/strings.xml
@@ -20,6 +20,8 @@
     <string name="value_button_8">8</string>
     <string name="value_button_9">9</string>
     <string name="value_button_0">0</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 
 </resources>
